+++ CinemaClub/Assets.xcassets/Add.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "profile.png",
+      "filename" : "Add.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
