+++ CinemaClub/Assets.xcassets/discover.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "profile.png",
+      "filename" : "discover.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
