+++ CinemaClub/Assets.xcassets/heart.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "profile.png",
+      "filename" : "heart.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
