+++ CinemaClub/Assets.xcassets/home.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "profile.png",
+      "filename" : "home.svg",
       "idiom" : "universal",
       "scale" : "1x"
     },
