--- CinemaClub/SwiftUIView.swift
@@ -1,19 +1,0 @@
-//
-//  SwiftUIView.swift
-//  CinemaClub
-//
-//  Created by Мария Степанчук on 04.06.2024.
-//
-
-import SwiftUI
-
-struct SwiftUIView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    SwiftUIView()
-}
-
